Erick

+++ src/main/java/com/turings/backend/model/Categoria.java
@@ -10,7 +10,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private  int id_categoria;
+    private  Long id_categoria;
 
     private String categoria;
 
@@ -24,11 +24,11 @@ public class Categoria {
         this.categoria = categoria;
     }
 
-    public int getId_categoria() {
+    public Long getId_categoria() {
         return id_categoria;
     }
 
-    public void setId_categoria(int id_categoria) {
+    public void setId_categoria(Long id_categoria) {
         this.id_categoria = id_categoria;
     }
